api/app/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
openai>=1.40.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
python-docx==1.1.2
pdfminer.six==20231228
reportlab==4.2.2
